pyproject.toml
[tool.poetry]
name = "pybotchi"
version = "3.0.0"
description = "A deterministic, intent-based AI agent builder."
authors = ["Alexie (Boyong) Madolid <madolid.alexie@gmail.com>"]
readme = "README.md"
repository = "https://github.com/amadolid/pybotchi"
packages = [{ include = "pybotchi" }]

[tool.poetry.dependencies]
python = "<4.0,>=3.12.0"
pydantic = ">=2.0"
orjson = ">=3.9.11"

# langchain-core dependency to be removed
# Should be optional
langchain-core = ">=0.3.15"

# Common optional
datamodel-code-generator = { version = ">=0.31.2", optional = true }

# MCP optional
mcp = { version = ">=1.15.0", optional = true }

# Server optional
click = { version = ">=8.3.1", optional = true }
grpcio = { version = ">=1.76.0", optional = true }
grpcio-tools = { version = ">=1.76.0", optional = true }

[tool.poetry.group.dev.dependencies]
mypy = "1.16.1"
python-dotenv = "1.1.1"
black = "25.1.0"
pre-commit = "4.2.0"
pytest = "8.4.1"
flake8 = "7.3.0"
flake8-import-order = "0.19.2"
flake8-docstrings = "1.7.0"
flake8-comprehensions = "3.16.0"
flake8-bugbear = "24.12.12"
flake8-annotations = "3.1.1"
flake8_simplify = "0.22.0"
pep8-naming = "0.15.1"
types-protobuf = "6.32.1.20251105"

# for examples
langchain-openai = ">=0.3.15"
fastapi = ">=0.109.0"

[tool.poetry.extras]
mcp = ["mcp", "datamodel-code-generator"]
grpc = ["click", "grpcio", "grpcio-tools", "datamodel-code-generator"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pybotchi-grpc = "pybotchi.grpc.cli:main"
pybotchi-grpc-compile = "pybotchi.grpc.cli:compile"
